+++ api/src/identity-access-management/infrastructure/repositories/reset-password-demand.repository.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../../db/knex-database-connection.js';
-import { PasswordResetDemandNotFoundError } from '../../../../lib/domain/errors.js';
 import { ResetPasswordDemand } from '../../../../lib/infrastructure/orm-models/ResetPasswordDemand.js';
+import { PasswordResetDemandNotFoundError } from '../../domain/errors.js';
 import { ResetPasswordDemand as ResetPasswordDemandModel } from '../../domain/models/ResetPasswordDemand.js';
 
 const RESET_PASSWORD_DEMANDS_TABLE_NAME = 'reset-password-demands';
